The BIG BIG commit !!!
git-svn-id: https://subversion.biologie.ens.fr/svn/leburon/trunk/eoulsan@1404 5a62d92b-e016-0410-9103-b89fe532bb34

+++ src/main/java/fr/ens/transcriptome/eoulsan/programs/mapping/local/FilterReadsLocal.java
@@ -142,7 +142,7 @@ public class FilterReadsLocal {
    * Set the threshold for the filter
    * @param threshold the threshold
    */
-  public void setQualityThreshold(final int threshold) {
+  public void setQualityThreshold(final double threshold) {
 
     this.qualityThreshold = threshold;
   }
